+++ modules/Multiplane/themes/rljbase/bootstrap.php
@@ -23,7 +23,7 @@ $this->set('multiplane.assets.bottom', [
 
 // extend lexy parser for custom image resizing
 $this->renderer->extend(function($content){ // returns relative url of scaled logo
-    return preg_replace('/(\s*)@logo\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".$app->retrieve("multiplane/lexy/logo/width", 200)."&h=".$app->retrieve("multiplane/lexy/logo/height", 200)."&q=".$app->retrieve("multiplane/lexy/logo/quality", 80); ?>', $content);
+    return preg_replace('/(\s*)@logo\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".mp()->get("lexy/logo/width", 200)."&h=".mp()->get("lexy/logo/height", 200)."&q=".mp()->get("lexy/logo/quality", 80); ?>', $content);
 });
 
 $this->renderer->extend(function($content) { // returns relative url of image
@@ -31,13 +31,13 @@ $this->renderer->extend(function($content) { // returns relative url of image
 });
 
 $this->renderer->extend(function($content){ // returns relative url of scaled image (thumbnail)
-    return preg_replace('/(\s*)@thumbnail\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".$app->retrieve("multiplane/lexy/thumbnail/width", 100)."&h=".$app->retrieve("multiplane/lexy/thumbnail/height", 100)."&q=".$app->retrieve("multiplane/lexy/thumbnail/quality", 70)."&m=".$app->retrieve("multiplane/lexy/thumbnail/mode", "thumbnail"); ?>', $content);
+    return preg_replace('/(\s*)@thumbnail\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".mp()->get("lexy/thumbnail/width", 100)."&h=".mp()->get("lexy/thumbnail/height", 100)."&q=".mp()->get("lexy/thumbnail/quality", 70)."&m=".mp()->get("lexy/thumbnail/method", "thumbnail"); ?>', $content);
 });
 
 $this->renderer->extend(function($content){ // returns relative url of scaled image (image)
-    return preg_replace('/(\s*)@image\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".$app->retrieve("multiplane/lexy/image/width", 800)."&h=".$app->retrieve("multiplane/lexy/image/height", 800)."&q=".$app->retrieve("multiplane/lexy/image/quality", 80)."&m=".$app->retrieve("multiplane/lexy/image/mode", "bestFit"); ?>', $content);
+    return preg_replace('/(\s*)@image\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".mp()->get("lexy/image/width", 800)."&h=".mp()->get("lexy/image/height", 800)."&q=".mp()->get("lexy/image/quality", 80)."&m=".mp()->get("lexy/image/method", "bestFit"); ?>', $content);
 });
 
 $this->renderer->extend(function($content){ // returns relative url of scaled image (headerimage)
-    return preg_replace('/(\s*)@headerimage\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".$app->retrieve("multiplane/lexy/headerimage/width", 800)."&h=".$app->retrieve("multiplane/lexy/headerimage/height", 200)."&q=".$app->retrieve("multiplane/lexy/headerimage/quality", 80)."&m=".$app->retrieve("multiplane/lexy/headerimage/mode", "thumbnail"); ?>', $content);
+    return preg_replace('/(\s*)@headerimage\((.+?)\)/', '$1<?php echo MP_BASE_URL."/getImage?src=".urlencode($2)."&w=".mp()->get("lexy/headerimage/width",800)."&h=".mp()->get("lexy/headerimage/height", 200)."&q=".mp()->get("lexy/headerimage/quality", 80)."&m=".mp()->get("lexy/headerimage/method", "thumbnail"); ?>', $content);
 });