+++ modules/Multiplane/themes/rljbase/post.php
@@ -1,13 +1,11 @@
 <?php
 // default view, if page type detection is "type" and the current type is "post"
 
-$width  = $app->retrieve('multiplane/lexy/headerimage/width', 800)  . 'px';
-$height = $app->retrieve('multiplane/lexy/headerimage/height', 200) . 'px';
+$width  = mp()->get('lexy/headerimage/width', 800)  . 'px';
+$height = mp()->get('lexy/headerimage/height', 200) . 'px';
 ?>
 
         <main id="main">
-
-            @render('views:partials/breadcrumbs.php', ['page' => $page])
 
             @if(!empty($page['image']))
             <img class="featured_image" src="@headerimage($page['image']['_id'])" alt="{{ $page['image']['title'] ?? 'image' }}" width="{{ $width }}" height="{{ $height }}" />